Rename FSMSpectrumImage to SpectrumContainerImage

+++ Modules/M2aiaCore/include/m2CoreCommon.h
@@ -178,7 +178,7 @@ namespace m2
                                                                 {"Double", 1}};
 
   using DisplayImagePixelType = float;
-  using NormImagePixelType = double;
+  using NormImagePixelType = float;
   using IndexType = unsigned int;
   using WorldCoordinateType = float;
   using ShiftImageType = int;
